+++ crates/api-structs/Cargo.toml
@@ -6,16 +6,22 @@ license-file.workspace = true
 
 [features]
 schema = ["dep:utoipa"]
-serde = ["dep:serde", "dep:serde_json"]
+serde = [] # keep it simple instead of adding conditional deps ["dep:serde", "dep:serde_json"]
 partial = []
 eq = []
 
 [dependencies]
 core-metastore = { path = "../core-metastore" }
-serde = { workspace = true, optional = true, features = ["derive"] }
-serde_json = { workspace = true, optional = true }
+
+arrow = { version = "55.1.0" }
+arrow-array = { version = "55.1.0" }
+
+indexmap = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
 chrono = { workspace = true }
 utoipa = { workspace = true, optional = true }
+snafu = { workspace = true }
 
 [lints]
 workspace = true
