+++ crates/api-structs/src/lib.rs
@@ -1,5 +1,16 @@
 pub mod auth;
 pub mod query;
+pub mod query_context;
 pub mod volumes;
+pub mod result_set;
+pub mod column_info;
+
 pub mod derive_features;
 
+// import with file name to avoid struct names conflicts
+// pub use auth::*;
+// pub use query::*;
+// pub use query_context::*;
+// pub use volumes::*;
+// pub use result_set::*;
+// pub use column_info::*;
