+++ crates/core-executor/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 license-file.workspace = true
 
 [dependencies]
+api-structs = { path = "../api-structs", features = [] }
 core-utils = { path = "../core-utils" }
 core-metastore = { path = "../core-metastore" }
 df-builtins = { path = "../df-builtins" }
@@ -32,8 +33,6 @@ iceberg-s3tables-catalog = { workspace = true }
 object_store = { workspace = true }
 regex = { workspace = true }
 snafu = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
 
 strum = { workspace = true }
 tracing = { workspace = true }
