+++ crates/core-executor/src/models.rs
@@ -1,7 +1,9 @@
 use datafusion::arrow::array::RecordBatch;
 use datafusion::arrow::datatypes::{DataType, Field, TimeUnit};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
+use api_structs::column_info::ColumnInfo as ColumnInfo_;
+
+// to keep existing imports unchanged
+pub type ColumnInfo = ColumnInfo_;
 
 #[derive(Debug)]
 pub struct QueryResultData {
@@ -9,133 +11,6 @@ pub struct QueryResultData {
     pub columns_info: Vec<ColumnInfo>,
     // query_id is QueryRecordId, but we won't add dependency on history crate here
     pub query_id: i64,
-}
-
-// TODO: We should not have serde dependency here
-// Instead it should be in api-snowflake-rest
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ColumnInfo {
-    pub name: String,
-    pub database: String,
-    pub schema: String,
-    pub table: String,
-    pub nullable: bool,
-    pub r#type: String,
-    pub byte_length: Option<i32>,
-    pub length: Option<i32>,
-    pub scale: Option<i32>,
-    pub precision: Option<i32>,
-    pub collation: Option<String>,
-}
-
-impl ColumnInfo {
-    #[must_use]
-    pub fn to_metadata(&self) -> HashMap<String, String> {
-        let mut metadata = HashMap::new();
-        metadata.insert("logicalType".to_string(), self.r#type.to_uppercase());
-        metadata.insert(
-            "precision".to_string(),
-            self.precision.unwrap_or(38).to_string(),
-        );
-        metadata.insert("scale".to_string(), self.scale.unwrap_or(0).to_string());
-        metadata.insert(
-            "charLength".to_string(),
-            self.length.unwrap_or(0).to_string(),
-        );
-        metadata
-    }
-
-    #[must_use]
-    pub fn from_batch(records: &[RecordBatch]) -> Vec<Self> {
-        let mut column_infos = Vec::new();
-
-        if records.is_empty() {
-            return column_infos;
-        }
-        for field in records[0].schema().fields() {
-            column_infos.push(Self::from_field(field));
-        }
-        column_infos
-    }
-
-    #[must_use]
-    pub fn from_field(field: &Field) -> Self {
-        let mut column_info = Self {
-            name: field.name().clone(),
-            database: String::new(), // TODO
-            schema: String::new(),   // TODO
-            table: String::new(),    // TODO
-            nullable: field.is_nullable(),
-            r#type: field.data_type().to_string(),
-            byte_length: None,
-            length: None,
-            scale: None,
-            precision: None,
-            collation: None,
-        };
-
-        match field.data_type() {
-            DataType::Int8
-            | DataType::Int16
-            | DataType::Int32
-            | DataType::Int64
-            | DataType::UInt8
-            | DataType::UInt16
-            | DataType::UInt32
-            | DataType::UInt64 => {
-                column_info.r#type = "fixed".to_string();
-                column_info.precision = Some(38);
-                column_info.scale = Some(0);
-            }
-            DataType::Float16 | DataType::Float32 | DataType::Float64 => {
-                column_info.r#type = "real".to_string();
-                column_info.precision = Some(38);
-                column_info.scale = Some(16);
-            }
-            DataType::Decimal128(precision, scale) | DataType::Decimal256(precision, scale) => {
-                column_info.r#type = "fixed".to_string();
-                column_info.precision = Some(i32::from(*precision));
-                column_info.scale = Some(i32::from(*scale));
-            }
-            DataType::Boolean => {
-                column_info.r#type = "boolean".to_string();
-            }
-            // Varchar, Char, Utf8
-            DataType::Utf8 => {
-                column_info.r#type = "text".to_string();
-                column_info.byte_length = Some(16_777_216);
-                column_info.length = Some(16_777_216);
-            }
-            DataType::Time32(_) | DataType::Time64(_) => {
-                column_info.r#type = "time".to_string();
-                column_info.precision = Some(0);
-                column_info.scale = Some(9);
-            }
-            DataType::Date32 | DataType::Date64 => {
-                column_info.r#type = "date".to_string();
-            }
-            DataType::Timestamp(unit, _) => {
-                column_info.r#type = "timestamp_ntz".to_string();
-                column_info.precision = Some(0);
-                let scale = match unit {
-                    TimeUnit::Second => 0,
-                    TimeUnit::Millisecond => 3,
-                    TimeUnit::Microsecond => 6,
-                    TimeUnit::Nanosecond => 9,
-                };
-                column_info.scale = Some(scale);
-            }
-            DataType::Binary | DataType::BinaryView => {
-                column_info.r#type = "binary".to_string();
-                column_info.byte_length = Some(8_388_608);
-                column_info.length = Some(8_388_608);
-            }
-            _ => {
-                column_info.r#type = "text".to_string();
-            }
-        }
-        column_info
-    }
 }
 
 #[cfg(test)]
